Switched community agreement over to about page
