+++ frontend/src/app/profile/profile-editor/profile-editor.component.html
@@ -97,19 +97,6 @@
 </mat-card>
 
 <mat-card *ngIf="profile.id" appearance="outlined">
-  <mat-card-title>Community Agreement</mat-card-title>
-  <mat-card-actions class="token">
-    <button
-      class="agreement-button"
-      mat-stroked-button
-      type="button"
-      (click)="openAgreementDialog()">
-      View Agreement
-    </button>
-  </mat-card-actions>
-</mat-card>
-
-<mat-card *ngIf="profile.id" appearance="outlined">
   <mat-card-actions class="token">
     <button
       class="display"
